ts_to_py_codes/langfuse/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
langfuse>=1.0.0
pydantic>=2.4.2
typing-extensions>=4.8.0 